% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coefCIN}
\alias{coefCIN}
\title{WID-CIN Coefficients}
\format{
## `coefCIN`
A named numeric with 5,000 coefficients. Names indicate corresponding CpGs. Weights were based on training using the entire Discovery Set.
}
\usage{
coefCIN
}
\description{
WID-CIN Coefficients
}
\keyword{datasets}
