% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeIndex.R
\name{computeIndex}
\alias{computeIndex}
\title{Basic index function}
\usage{
computeIndex(beta, w, scaling, name = "", invert = T)
}
\arguments{
\item{beta}{Beta-valued matrix with rownames labeling CpGs and columns labeling samples for which an index is desired}

\item{w}{Coefficients/weights}

\item{scaling}{Scaling: this is data the index is centered around}

\item{name}{index name (optional)}

\item{invert}{Should index be inverted? This will depend on the index and how it was trained.}
}
\value{
Scaled index
}
\description{
Basic index function
}
