% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scalingCIN}
\alias{scalingCIN}
\title{WID-CIN Scaling}
\format{
## `scalingCIN`
Index scaling (n=638), based on values of the index in the Discovery set.
}
\usage{
scalingCIN
}
\description{
WID-CIN Scaling
}
\keyword{datasets}
