% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WID_CIN.R
\name{WID_CIN}
\alias{WID_CIN}
\title{WID_CIN index}
\usage{
WID_CIN(beta)
}
\arguments{
\item{beta}{Methylation beta matrix object (Illumina Methylation EPIC; or 450K array)}
}
\value{
Returns a list with CIN index (training and discovery version, and island versus open sea CpGs)
}
\description{
THE WID_CIN function uses a methylation beta matrix and returns the WID-CIN-index.
Authors: Barrett J., Herzog C. et al., 2021
}
